refactor new logging API
