+++ platform-operator/controllers/verrazzano/component/istio/istio_install.go
@@ -5,6 +5,7 @@ package istio
 
 import (
 	"context"
+	vzlog "github.com/verrazzano/verrazzano/pkg/log/vzlog"
 	"io/ioutil"
 	"os"
 	"path/filepath"
@@ -17,7 +18,6 @@ import (
 	"github.com/verrazzano/verrazzano/platform-operator/internal/config"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/istio"
 
-	vzlog "github.com/verrazzano/verrazzano/pkg/log/progress"
 	istiosec "istio.io/api/security/v1beta1"
 	istioclisec "istio.io/client-go/pkg/apis/security/v1beta1"
 	appsv1 "k8s.io/api/apps/v1"
