+++ platform-operator/controllers/verrazzano/component/spi/component_context_test.go
@@ -4,11 +4,10 @@ package spi
 
 import (
 	"github.com/stretchr/testify/assert"
-	vzlog "github.com/verrazzano/verrazzano/pkg/log/progress"
+	vzlog "github.com/verrazzano/verrazzano/pkg/log/vzlog"
 	clustersv1alpha1 "github.com/verrazzano/verrazzano/platform-operator/apis/clusters/v1alpha1"
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/config"
-	"go.uber.org/zap"
 	"io/ioutil"
 	istioclinet "istio.io/client-go/pkg/apis/networking/v1alpha3"
 	istioclisec "istio.io/client-go/pkg/apis/security/v1beta1"
@@ -140,7 +139,7 @@ func TestContextProfilesMerge(t *testing.T) {
 			assert.NotNil(expectedVZ)
 
 			// Create the context with the effective CR
-			log := vzlog.EnsureLogContext("test").EnsureLogger("test", zap.S(), zap.S())
+			log := vzlog.DefaultLogger()
 			context, err := NewContext(log, fake.NewFakeClientWithScheme(testScheme), &test.actualCR, false)
 			// Assert the error expectations
 			if test.expectedErr {
